move icon type to seperate folder
