+++ src/components/fundamentals/icons/bell-noti-icon.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import IconProps from "./icon-type"
+import IconProps from "./types/icon-type"
 
 type IBellNotiIconProps = IconProps & {
   accentColor?: string
