+++ src/components/fundamentals/icons/dollar-sign-icon.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import IconProps from "./icon-type"
+import IconProps from "./types/icon-type"
 
 const DollarSignIcon: React.FC<IconProps> = ({
   size = "24px",
